[unwinder] Fix 32 bit FP unwinding
...by no longer skipping this unwinder when we see a 32 bit register
set.

The PC validation logic is now contingent on being able to load CFI for
the module containing the next PC value - 32 bit modules might not have
any CFI, but FP unwinding succeeded. We don't want to prematurely skip
the FP unwinder in this case since it's difficult to validate the PC.
The downside is we might report a bogus frame in the case the PC is
invalid.

Bug: 395606674
Bug: 416040479
Change-Id: I6a6a6964388294c57c5101620f6ab9da5552152e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1351564
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Reviewed-by: James Robinson <jamesr@google.com>