[rust] Fix mutability of ioctl args in test
libc::ioctl's third parameter is `...`. Rust doesn't provide any
guarantees on arguments passed to this parameter. The test was passing a
shared ref to dm_ioctl, while the function was modifying the struct.
This triggered undefined behavior. When compiled in release mode, the
compiler assumed the function could not modify the value and so
optimized the assertions. It could see from usage that the flags were
not modified and so pre-computed the failure, while allowing the logs to
still read from the values passed to the assertion, leading to confusing
errors like:

assertion `left == right` failed
  left: 66
 right: 66

Fixes: 441804545
Test: fx test
//src/starnix/tests/syscalls/rust:starnix_syscalls_rust_tests in release
profile

Change-Id: I5472556fb8338f95fa230de4f2d5aba30e18584f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1356898
Fuchsia-Auto-Submit: Paul Faria <paulfaria@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Paul Faria <paulfaria@google.com>