[starnix] Reference MemoryManager directly in "/proc/*/mem" files
When "/proc/<pid>/mem" is opened, the resulting FD should be linked
to the target task's current address-space, rather than to the task
itself. This creates a security risk on exec(), since existing FDs
may be used to access the task's new post-exec address-space, even
if the task runs with different credentials (e.g. it is suid-root
or has a new SELinux Security Context).

Instances of "/proc/<pid>/mem" now take a weak reference to the
MemoryManager for the Task, becoming inaccessible as soon as the
last reference is dropped. This requires that the Task's reference
to the MemoryManager be (briefly) locked whenever a caller wishes
to reference it.

On exec() the MemoryManager is completely replaced, in four steps:
1. The old MM is locked and the user VMAR & mappings removed.
2. The new user VMAR and MM is created.
3. The old mappings are dropped.
4. The new MM is assigned into the Task structure.

Bug: 437995606
Change-Id: I36e4d3549035c5fe78f5ce0edbc534b596ed5a1c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1343804
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Wez <wez@google.com>
Reviewed-by: Maria Glukhova <mariagl@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>