[utf-utils][bazel][idk] Use bazel2gn
Allow `friends` to be passed to `idk_cc_source_library()` for
bazel2gn to use.

Note that while there is no //zircon/system/ulib/zx Bazel target, this
is okay because utf-utils is currently only built in Bazel for on the
host platform. The reference is necessary for bazel2gn to generate the
correct GN definition.

Bug: 428229472
Test: fx build && fx bazel build --platforms=//build/bazel/platforms:host //sdk/lib/utf-utils:utf-utils
Change-Id: I1e671d6037a95cef1f907b3e3afdac70d33498a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1327290
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>