[assembly] Add TestRamdisk mode
This mode will allow a test to provide a "ramdisk" which may not
be a properly formed ZBI. Assembly should not attempt to modify this
image, but still shove it into the product bundle as the "ZBI". It
will get booted by emulators as the initrd.

Test: fx build //src/tests/assembly/test_ramdisk
Bug: 439032069
Change-Id: I06d90204fe90344ac7c107855090919ec2180058
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1352052
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Aidan Wolter <awolter@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>