Reland "[starnix] Reference MemoryManager directly in "/proc/*/mem" files"
This is a reland of commit fbd1d796cedff9007e6c1618bd0434e0e87743dc,
with additional logic to reset the user address-space VMAR handle, and
accommodate Zircon mapping queries made via lingering references to the
old MM while is in that state.

When "/proc/<pid>/mem" is opened, the resulting FD should be linked to
the target task's current address-space, rather than to the task itself.
This creates a security risk on exec(), since existing FDs may be used
to access the task's new post-exec address-space, even if the task runs
with different credentials (e.g. it is suid-root or has a new SELinux
Security Context).

Instances of "/proc/<pid>/mem" now take a weak reference to the
MemoryManager for the Task, becoming inaccessible as soon as the last
reference is dropped. This requires that the Task's reference to the
MemoryManager be (briefly) locked whenever a caller wishes to reference
it.

On exec() the MemoryManager is completely replaced, in four steps:
1. The old MM is locked and the user VMAR & mappings removed.
2. The new user VMAR and MM is created.
3. The old mappings are dropped.
4. The new MM is assigned into the Task structure.

Bug: 437995606
Change-Id: I9c5d5b7f3cdeba62eb0cec94ba1364ac11c65a8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1358411
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Wez <wez@google.com>