[kernel][vm] Manually specify zero overwrites
Adds an additional parameter to CompleteAddPageLocked where the caller
can specify whether they know if any visible parent content was the
true zero page or not. This allows CompleteAddPageLocked to correctly
specify whether the unmap operation is Unmap or UnmapZeroPage, as
previously it did not have enough information to state this correctly.

Bug: 441409277
Change-Id: I602c2f39ff67984dbe6517c853258a52d169dc97
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1357264
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>