[unwinder] Recognize S augmentation string
...which is produced by the .cfi_signal_frame directive. This indicates
that the frame we are currently constructing explicitly sets PC itself,
and we should not override it with the return address register and
symbolizers should not modify the PC for symbolization to avoid off by
one errors.

This adds a new field to each Frame object, which is populated by the
S augmentation. This flag isn't used yet but a followup will add 32 bit
support to the Sigreturn unwinder which will be triggered this this
flag.

Bug: 395606674
Bug: 416040479
Change-Id: I6a6a6964bd47d8169061b61c32510ea88905b914
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1352945
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>