[unwinder] Support 32 bit sigreturn unwinding
This probes for the 32 bit equivalent of the already supported 64 bit
siginfo struct and __kernel_rt_sigreturn function.

This change also adds probing for sigreturn instructions so we can
identify signal frames even when .cfi_signal_frame is not specified or
we do not have any CFI data to use.

The 32 bit implementation in particular is extremely difficult to test
within our current testing scheme, due to the tests running the unwinder
in the same process that is being unwound. It is impractical (and not
entirely useful) to test 32 bit binaries on the host today, and on
Fuchsia we need Starnix to run them.

Tests may be added later in the form of zxdb_e2e_tests or adding
additional testing here that better emulates a real environment.

Bug: 395606674
Bug: 416040479
Change-Id: I6a6a6964561e242ed497ee7468cdacba861f8ed6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1351146
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>