[kernel][timers] Blunt some of the sharp AV7T edges.
Update the Armv7MmioTimer implementation to not hold any locks during
its IrqHandler callbacks.  Also borrow some techniques from the core
Timer implementation to preserve the properties of methods like
`CancelTimer` to guarantee that after a call to `CancelTimer`, the
existing handler has already run to completion, or it has been
successfully canceled and will not run at all (provided that no new
calls are made to SetTimer).

Add usage docs to the AV7T implementation which describes how the
timer is meant to be used, and documents these guarantees.  In
particular, document the fact that it is now safe for a IrqHandler for
the timer to call back into the timer instance methods themselves.

Also Update the AV7T timer implementation and shell commands to
prevent the shell commands from accidentally reprogramming a timer
instance which is actively being used by something like the
SuspendWakeupTimer.  Shell commands can still inspect timer state.

Finally, update the docs for the one current user of the AV7T timer,
the SuspendWakeupTimer, to reflect the fact that it is now safe for
the timer's registered callback to call into its SuspendWakeupTimer
instance, while acknowledging the fact that there is probably little
practical reason to ever want to do this (but at least it is safe
now!).

Change-Id: I8f252b171088cb908f54475d97e13a02a181216c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1356895
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: John Grossman <johngro@google.com>