[bazel_sdk] add retry logic to run_buildifier
We are seeing flakes in our builder which reports that files in the SDK
are not formatted correctly. We put in some debug logging that prints
out the contents of the failed files when this happens but they do not
have errors which leads me to believe that this is a flake in
buildifier. This retry logic will help us determine if this is a flake
in our build determinism or a flake in buildifier.

Change-Id: I9b4f2201a57b3dea990bd074775397e0153cb0f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1357454
Reviewed-by: Darren Chan <chandarren@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>