[starnix] Ambient capabilities properly updated on calls to capset()
Previously, if a call to capset() were to reduce either the permitted or
inherited capability sets for a thread, then the ambient set would stay
the same.  This CL modifies capset() to maintain the invariant that the
ambient capability set is always a subset of the permitted and inherited
set.

Bug: 371550616
Change-Id: I7d5a58202fbaebe6554efb3b11a7311f11374087
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1356224
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Tomás Dougan <tomasdougan@google.com>