[debugger] Initialize IPC versions during integration tests
I noticed this from a test failure while trying to land
https://fxrev.dev/1352364.

The debug_agent integration tests were subtly broken by a change to the
StackFrame record, which will soon look at the version field during
serialization. These tests were not initializing the IPC version to a
known version, so the serialization routine saw IPC version
0. This directly conflicts with the LocalStreamBuffer implementation,
which supplies the current protocol version directly to the
deserialization routines, which means the StackFrame records expect to
be able to deserialize the new fields, when they haven't been
serialized.

In effect, this simulated an uncaught IPC negotiation failure where things
proceed to go completely off the rails, which was not easy to debug.

This also adds an assert to the debug_ipc notification path that asserts
if we try to send a notification when the IPC hasn't been initialized
yet to make this easier to catch in the future.

Change-Id: I6a6a6964f85c5fda04df61edee509ce59775fb52
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1360332
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>