[ffx][trace] Removing trace daemon protocol
This removes the tracing protocol from the ffx daemon and changes the
`ffx trace` subtool to use a direct connection to the target device.

This implementation is backwards compatible with devices that do not
have the /core/trace_manager/trace_session_manager component for
foreground tracing only.

`ffx trace` will first attempt to use the SessionManagerProxy, at
/core/trace_manager/trace_session_manager, then if it is not there, fall
back to the ProvisionerProxy, at /core/trace_manager.

The communication in both cases is direct to the device; no daemon is
involved. As a side effect, background tracing using the
ProvisionerProxy is not supported since there is no process to hold open
the connection.

As a result, a trace can be started and no connection is maintained to
the device so it can truly idle. Then `ffx trace stop` can be used to
copy off the trace data.

Bug: 431753760

Change-Id: I351a83e6397713a4bef8200f2174b1ab04235931
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1349444
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Fadi Meawad <fmeawad@google.com>