[sestarnix] Add audit framework config
Implemented the NETLINK_AUDIT socket ops. The requests coming from
userspace are processed by the `AuditNetlinkClient`, which is the
interface between the socket and the `AuditLogger`. It contains an
optional netlink response, which will be sent if the audit daemon
reads its socket. We can assume that the audit daemon will wait for
a response from the `AuditLogger` and will block. When the audit
daemon tries to set or get the audit status of the logger, its task
must have the capabilities CAP_AUDIT_* needed or that specific request.

The `AuditLogger` prints all the messages to the syslog, but can
set its configuration when requested by an `AuditNetlinkClient`,
which will influence the behavior of the logger. The `AuditLogger` can
be disabled at boot by adding "audit=0" in the kernel command line.

Bug: 411810652
Change-Id: Ic6acfc832824fade61e14d7842b8e1710c45963e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1342585
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Rareş Constantin <rarescon@google.com>