[sestarnix] Add sink-logging to AuditLogger
The `AuditLogger` is the component responsible with writing the
audit messages back to the socket that is registered as the audit
daemon. The `AuditLogger` contains a deque which represents the backlog
for audit messages. It will notify the `AuditClient` registered as the
audit daemon when a new audit message is received. If there is no audit
daemon available, the logger prints the audit messages to syslog.

If the audit daemon is waiting for a response, the `AuditNetlinkClient`
will first send the reply to that request if present. Otherwise, it
will send an audit message from the backlog.

When the backlog limit is reached, the userspace threads trying to
audit an event will drop the message and will print this to syslog.

Bug: 411810652
Change-Id: I68082e980d05ff34469aabc92217d4d63dbb6b2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1350184
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Rareş Constantin <rarescon@google.com>