[starnix] Add kernel cmdline opts parsing
The Kernel may be configured using the kernel command line at runtime.
These options can be applied to subsystems such as the `AuditLogger`.

Each subsystems that needs to configure itself based on the command line
may use the `cmdline_args_iter()` function, which returns an `Iterator`
that contains the arguments as name and value.

Bug: 440087162
Change-Id: Icb29c39d0d09cb705e80abb6a2d0f2779e815885
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1359737
Commit-Queue: Rareş Constantin <rarescon@google.com>
Reviewed-by: Wez <wez@google.com>