[kernel][vm] Remove lock on prng accessor
As per the comment on the declaration of aslr_prng_, the object is
thread safe and does not need to be guarded by the lock, making this
lock requirement redundant.

Note that the origin of the lock requirement is that in the past the
ASLR configuration utilized non-const variables, and as such what is
now the `is_aslr_enabled` accessor required the lock.

Change-Id: I48734927a7ab94c1a4f8ebb9bb55f852cfe0e6d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1360431
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>