[bazel][idk] Rename sdk_host_tool() to idk_host_tool()
For consistency with the other IDK macros.

Also drop "_sdk" from the `name` argument in instances of the macro and
update bazel2gn to append it for compatibility with the GN template,
which does not automatically add the suffix.

Remove `kwargs` because there will be no underlying target to accept the
additional args.

Bug: 442025401
Change-Id: I0615146cc25241b854677420de0271c03ab136af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1357712
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>