[bazel][idk] Require all atoms are in an allowlist
Define a function to return the allowlist corresponding to an atom's
properties and use it to add a dependency on the appropriate allowlist
in idk_atom(). This ensures that all atoms are in an allowlist.

Keep the dependency in the underlying cc_library() target to handle
cases where the macro is used but there is no deps on the atom target.

Make `atom_build_deps` a required parameter to avoid "trying to mutate
a frozen list value" errors when the default value is used.

Bug: 417305295
Change-Id: Id6f3fa3ba395705dc677236c099ade385c71787d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1357772
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>