[tools] Add custom analytics to g-review
Adds custom analytics to greview to track impact.
Creates an anonymized identifier for the current CL,
to track if users find it useful to re-run the tool
on the same CL

Asks the user if it's ok to append the label '[g-review]'
to the commit message, to help track bug speed through review.

Additionally records if g-review gave lgtm, how many changes it
suggested, and if the user found the review helpful.

Test: manually testing on Fuchsia CLs
 - Verified ~/.fuchsia/g-review is created and SALT is populated
 - Verified user response to helpfulness is tracked
 - Verified label added to commit message on user approval

FIX: 441985229

[g-review]

Change-Id: I4ae0a7830ec2a77d1b2f105cb4d5896dbd83949a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1356286
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: David Murphy <djmurphy@google.com>