[ffx][mdns_discovery]: fix file descriptor leak
When querying our recv loops, we have a HashMap of Ip -> Task that keeps
track of the different queries we are doing concurrenlty. The task
itself, had a strong reference (Rc) to the list of tasks, creating a
circular reference so that the tasks could not be deallocated or
dropped.

Switching query_recv_loop to a `Weak<Mutex<HashMap>>` as opposed to a
`Rc<Mutex<HashMap>>` allows the various Tasks to be dropped and the Udp
sockets for them to be cleaned up and Closed.

Tested: by running `fx serve` in one window (as it does discovery every
10 seconds) and running `watch "sudo netstat -tupan | grep ffx"` to see
how the udp descriptors are created on each query loop and subsequently
dropped.
Bug: b/440391178, b/440391178
Fixed: b/442884174
Change-Id: Id138469a6ab50109b863ed2d8ab42fcf6a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1360782
Reviewed-by: Steven Grady <slgrady@google.com>
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>