[fake-bti] Take ownership of paddrs
Rather than having CreateFakeBtiWithPaddrs retain a reference to the
paddrs it will use for pinning (and the client mutates the paddrs in the
span if needed), instead have the FakeBti take ownership of the paddrs
by taking a copy.

Adds fake_bti::SetPaddrs as a replacement for clients which need to
change the set of pinned paddrs after creation.

There is a slight behavioural change with SetPaddrs:  Whenever SetPaddrs
is called, the index into the paddr list is reset, so clients will start
getting paddrs from offset 0 for new pin requests.  No tests seemed to
rely on the old behaviour (aml-sdmmc-with-banjo-test worked around it),
so this seems fine, and it's easier to understand the behaviour now.

This refactoring is a prerequisite for creating a Rust wrapper for
FakeBti, since we don't have a way to express a lifetime for a pointer
that we pass into the C FFI interface for CreateFakeBtiWithPaddrs.  It
is also arguably less error-prone.

Bug: 442012466
Change-Id: I88c37d727d7f26c44dd7af5468c55965d8525601
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1360777
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
API-Review: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>