[battery] Revive Battery Manager
Before this CL, Battery Manager isn't used in products anymore once
the workstation is deprecated. In certain products the fake battery
driver is used directly with client bypassing the battery manager.

In this CL, Battery Manager is revised to connect to battery driver
using battery fidl and then forward the BatteryInfo to clients as a
bridge.

The wake_lease for Battery Simulator mode isn't fully covered yet and
will be added later if necessary.

Bug: b/422755268
Test: CQ
  fx test battery-cli-tests
  fx test battery-manager-tests
Change-Id: I567e8574ceb5d4c22b7698ef5d3f807a4496c80b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1351606
Commit-Queue: Hai Bi <bihai@google.com>
Reviewed-by: Michael Brunson <mbrunson@google.com>