[fxfs] Install volumes in-place instead of copying
As part of this change, we also remove the existing platform code as we
are going to update how volume installation works. The protocol needs to
be reworked to support writing sparse chunks, and needs to be accessible
at the filesystem level to facilitate creation of a new volume
(currently everything is tied to an existing volume, which is not ideal
since it requires we delete everything that may exist).

Bug: 415300916
Test: fx test fxfs-tests --test-filter="*install*"

Change-Id: I77e642d6a0854a8fc6f1023df0306e45b02a1594
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1319624
Reviewed-by: James Sullivan <jfsulliv@google.com>
API-Review: Chris Suter <csuter@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>