[kernel]][sched] Reorganize power domain lifetime
Reorganize the way PowerDomain instances are registered and shared with
individual CPUs.

Before this change, PowerDomainRegistry maintains a linked list of all
PowerDomain instances and each CPU is updated with a ref ptr to the most
recent version of the specific PowerDomain the CPU belongs to. This is
sufficient for making local power estimate calculations (e.g. updating
thread and CPU energy counters during reschedules) without additional
locking. However, it is insufficient for task placement calculations,
which may need to compare the energy costs across domains.

In this change, a new PowerDomainSet type is introduced that maintains
an array of ref ptrs to PowerDomain instances representing the complete
power domain configuration of the system. PowerDomainRegistry maintains
and updates the primary power domain set and each CPU is given a
complete local copy of the set whenever the primary set is updated. This
allows each CPU to perform all of the necessary calculations locally
without acquiring other CPU's queue locks or introducing commonly shared
locks for these operations.

This change also introduces direct usage of ulib/ffl for utilization and
processing rate quantities to simplify usage in the kernel, avoiding
manually wrapping/unwrapping these quantities outside of the library.
Static assertions are added to ensure that the equivalent kernel and
scheduler types stay in sync.

Bug: b/322206755
Bug: b/322207536
Change-Id: Ide1312ed80c4cdd2dbf6f0fd31c13454facf86d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1359936
Commit-Queue: Corey Tabaka <eieio@google.com>
Fuchsia-Auto-Submit: Corey Tabaka <eieio@google.com>
Reviewed-by: John Grossman <johngro@google.com>