[fshost] add extra slice to limit when using zxcrypt
When we launch zxcrypt as a driver, we added an extra slice to the limit
to account for the zxcrypt metadata. This adds the same behavior when
setting the limit with the new fvm component, which takes care of
zxcrypt internally.

Fixed: 443074520
Change-Id: Ic520b00fc5e807790575b86339e22ac09d5310ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1361567
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>