[trace2json] Add support for filtering events by name
This patch adds a `--pattern` command line argument to `trace2json` that
allows filtering trace events by their name. The pattern is a regular
expression that is matched against the event name. Events that do not
match any of the provided patterns are dropped. Scheduler and metadata
records are always retained.

The caller may specify `--pattern` multiple times. Events that match ANY
of the provided patterns will be retained.

This is useful for reducing the size of trace files and for focusing on
specific events of interest.

Bug: 437397276
Test: fx test trace2json_tests

Change-Id: Icb8dbc752076dc97985166150a59073bb2ba977a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1361756
Reviewed-by: Jia Li <lizhijia@google.com>
Commit-Queue: Chris Masone <cmasone@google.com>