[starnix] remote_uds: Return ECONNRESET on peer closed
When a FIDL read call on a remote unix domain socket returns a
PEER_CLOSED error, this indicates that the remote end of the socket
has been closed. The expected behavior for a socket in this state is
to return ECONNRESET on subsequent read attempts, as documented in
the unix(7) man page.

This change updates the error handling to correctly map the
zx::Status::PEER_CLOSED status to the ECONNRESET errno.

Reference: https://man7.org/linux/man-pages/man7/unix.7.html

Bug: 357665862
Test: fx test starnix-tests --test-filter "*remote_unix_domain_socket::tests::test_remote*"
Change-Id: Ifc5faa7286d4a3fc0b9d024cb012f73de28492d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1362399
Reviewed-by: Guocheng Wei <guochengwei@google.com>
Commit-Queue: Chris Masone <cmasone@google.com>