[fidl][rust] Add concurrency testing via loom
Loom can ensure that every interleaving of concurrent operations never
deadlocks and never causes UB. This CL introduces a new concurrency
module to the new bindings procol layer. It wraps the concurrency
primitives from the standard library and futures crate so that they have
the same API as the loom concurrency primitives. Adapting the existing
code to use the same API with and without loom allows us to compile the
protocol layer with loom types as a separate library without forking the
source code. Finally, this CL adds a new library and test target for
loom tests, ports the standard suite of tests to run on loom, and fixes
a bug that loom found with atomic orderings.

Fixes: 441700153
Change-Id: I555f651bc1eba3f5e3f0d6ee0f342db677652a54
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1357189
Commit-Queue: David Koloski <dkoloski@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>