[fdf][fidl] Add WaitForDriver to node controller
Adds a new fidl method that is a hanging get that returns when a driver
is started on a node (or under a composite parent node). Or returns
error if there is an error in matching, or starting a driver.

Bug: 441302490
Change-Id: I179c6ab63672cadaa6e762e605ff6081b841687a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1355384
API-Review: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>