[blobfs] Migrate to V10 supporting mixed merkle layouts.
Migrates v8 and v9 to v10 where an inode flag is used to determine the
mkerle format, and a superblock flag that determines which format to use
for writing by default. This allows us to prepare systems for migration
without actually changing layouts on disk. The migration will be
completed with a minor version bump that changes the superblock flag to
unify all future writes.

Both the superblock flag and the inode flag are 1 for the old version so
that when migration is finished the flag will be zeroed out and can be
reclaimined for future uses.

Bug: 415108303
Change-Id: I99a4c387553d87581ae4ae01f13eb67d38c9ea7b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1358032
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>