[trace2json] Use full match for event name filtering
The event name filtering in trace2json was using partial regular
expression matching. This was unintentional and could lead to surprising
results when a user provides a pattern that is a substring of multiple
event names, as more events than intended would be included in the
output.

This change corrects the filtering logic to use a full match against the
event name. This ensures that only events that match the entire pattern
are included in the output.

The existing tests have been updated to reflect this change, and a new
test case has been added to verify that exact matches work as expected.

Bug: 437397276
Test: fx test trace2json_tests
Change-Id: I54ddcf238a560883fa555185d4bf039f2469adce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1364340
Fuchsia-Auto-Submit: Chris Masone <cmasone@google.com>
Reviewed-by: Jia Li <lizhijia@google.com>
Commit-Queue: Chris Masone <cmasone@google.com>