[scenic] Ensure display mode is properly set.
fxr/1208811 introduced an issue where the display mode is never set if
the next display config is dropped instead of applied (for example, when
we attempt direct-to-display but need to fall back to GPU composition).

Now we remember the mode that needs to be applied.

Fixed: 402804098
Run-All-Tests: True
Change-Id: I4bf5d1ff0e80118ef9a5e863a4b065a659e454e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1252865
Commit-Queue: Josh Gargus <jjosh@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>