[kernel][arm64] Improve TopByteIgnoreTests.InstructionAbortNoTag
Improve TopByteIgnoreTests.InstructionAbortNoTag by adding a
regression test for https://fxbug.dev/441337397.

Also, in arm64_instruction_abort_handler, move the point at which we
strip the FAR tag to after we would panic if the exception were
EL1->EL1.  The idea is that in the case of a kernel panic, we should
preserve as much information as possible to assist in debugging.

Fixed: 441337397
Change-Id: I17392e0c62cbd3176ff5c7a341980f1f90d5f8ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1356954
Fuchsia-Auto-Submit: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>