[f2fs_reader] Finalize once, at the end.
I originally worried that allocations that exhaust the allocator's
cache of extents in image_builder_mode might cause issues because
the allocator borrows from power-of-two extents up to 2MiB and caps
it's cache of free extents thereafter at a capacity of 512.

This fear was unfounded because temporary_allocations are tracked
and so if we exhaust the first ~1GiB of disk in this way, we will
attempt to scan the allocator LSM tree, which is empty except for
unwritten extents and temporary allocations. This will correctly
refresh another 512 2MiB extents.

To quell my concerns, I tested this by setting
NUM_MAX_SIZED_EXTENTS_TO_KEEP to 1 and tracking the rebuild_strategy
inspect histogram. 6 calls and no failures so I've left finalize()
last and added some additional device reopen() calls to ensure we keep
the device read-only until we absolutely need to write.

Bug: b/373995442
Change-Id: Ice7d1017402a968f42aff698152afe5d1d169289
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1360432
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Aaron Drew <ripper@google.com>