[graphics][vulkan_loader] Add more logs for debugging flakes
From the last CL we can see that we are not getting the idle
signal. Add a helper function to see if we are connecting to
the directory at all or seeing a single file out of it.

Bug: 435953902
Change-Id: I6e8691696d3bbc0381570d2bd7b226de64c26be5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1365231
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>