[ffx] ffx discovery API refactor [reland]
(This is a reland of fxr/1357783, with the problematic timeouts in
bootloader and flash fixed.)

The target discovery API has been refactored to better support one-shot
queries, simplifying many common use cases and preparing for a future
caching layer. The primary Discovery::discover_devices() function is
now an async method that returns a Vec<TargetHandle>. The previous
stream-based functionality is preserved in a new discovery_stream()
function for callers that require continuous monitoring.

As part of this change, the notify_added and notify_removed methods
were removed from DiscoveryBuilder. Internally, discover_devices() now
consumes the discovery_stream() and uses a TargetSet helper to process
Added and Removed events, building a final, consistent list of targets
to return to the caller.

Test: since it is solely a refactor, we rely primarily on regression
tests

Change-Id: I6a6a69641e39730962fbf98ee679c17ef3739222
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1362363
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>