[remote-control] runner write to stdout without FDIO
Instead of using the `print!` macro to write the compatibility info,
grab the underlying fd handles to stdout and write from there. This
bypasses FDIO.

This fixes a panic that can happen when the logging buffers are full on
the target.

Tested: Used a component to generate a large amount of log spam and then
used `ffx` to connect to the target. Observed there was no stack dump in
the serial logs of the Target
Change-Id: I9a584e58ee2bc35fb1d78fd094b0bd626a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1364884
Commit-Queue: Colin Nelson <colnnelson@google.com>
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>