[dwc3] Initial Inspect addition to the dwc3 driver
This CL adds a minimal set of Inspect data tracking. The machinery is encapsulated in the dwc3-metrics files.

The basic model of the metrics collection is to gather the data outside of Inspect and to create lazy Inspect nodes that copy in the data when Inspect data is
requested. This is currently the best practice for memory-constrined systems.

Bug 435013351

Change-Id: I9366f5fc5abfeaab4f5112e5176e4ffeaf1b7349
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1361211
Reviewed-by: Sam Hansen <hansens@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: John Sasinowski <jsasinowski@google.com>