[bazel] assert directory outputs always specify tracked files
This adds a GN assert that the 'directory_outputs' of a bazel
action always specify tracked files.  It also removes the output
dir from the list of output files of the GN action, as it's not
a file.

Change-Id: Ifc48a1bbd2c29a9da4063ba9a727cbc0f35e979e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1365265
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>