[paver] Support MergeSuperAndUserdata
fuchsia.fshost.MergeSuperAndUserdata was added recently to have the GPT
component logically merge "super" and "userdata" into a single partition
when possible.  Update the paver to also look for this merged partition
when the config value is set.

Note that the majority of this change is mechanical refactoring to plumb
PaverConfig throughout the paver.  The main business logic changes are
in android.cc and moonflower.cc.

Bug: 428278682
Change-Id: I21313c9c3531a8b7f7d80696a9a7c51d694a30db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1362557
Reviewed-by: David Pursell <dpursell@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>