[fastboot] Fix availability of fuchsia.fshost.Recovery
The `fuchsia.fshost.Recovery` protocol is not available on all build
configurations, so we need to ensure the use declaration reflects this
correctly. Additionally the capability route was missing from the core
shard preventing fastboot from working on devices.

Bug: 377560660
Test: Verified `ffx component doctor fastboot-usb` on a device and that
the protocol is usable with a custom OEM command.

Change-Id: Ia92bf10a63a07f3abe757391880d191909faba23
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1364365
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: David Pursell <dpursell@google.com>