[early-boot-coverage] Only use fxfs images in coverage_verification_test
This removes the use of the global 'fxfs_blob' GN arg, so that the
test always runs with an fxfs image in the Fuchsia image used to
return the test results to the host.  The test itself runs before
storage is brought up, and doesn't need the fvm vs. fxfs distinction.

Change-Id: I414d5276456935c3f6a36f65fd25acb4a18b371b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1365535
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>