[ui-test-realm] Set fake display configuration.
This sets mutable configuration of the fake-display-stack-host
in ui-test-realm based tests. This makes it possible for tests
to define the display resolution / refresh rate. In the future
we'll be able to test multi-layer compositing when it's supported
in fake-display.

Test: //src/ui:tests
Bug: 129364
Change-Id: Ib46774d8a0cbcb052f28c80a1b35df9fb9040abd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/876863
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>