[reachability] Improve default route detection
The default route telemetry reporting was relying on the
has_default_v{4,6}_route fields from the interfaces watcher. These
fields are only set to true if the default route for the device is in
the main table. It is possible for an interface to have a default route
active in a separate table, so we should check through the routes to
determine whether there are any v4/v6 default routes present.

Fixed: 440616646
Change-Id: I019b1e5ae5f0c0155b92890d73cd72ca972d65fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1361557
Reviewed-by: Zoey Bush <zmbush@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>